requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
pkg-resources==0.0.0
requests==2.27.1
urllib3==1.26.12
websocket-client==1.3.1
